Bump version to 1.1.0
The rationale for increasing the minor version (instead of only the
patch level) is that we enforce a new version of the pulsectl dependency
with this update. With the new minor version we could (in theory) add
patch releases for 1.0.0, if someone needs to stick to the old pulsectl
version.